Another large commit (hopefully the last for now)
This gets the code to the point whereby it implements a fake camera using
the code from the emulator. The code required to start using the sensor
is the next piece of the puzzle, but at least now the framework is all
in place. The generated image isn't right (it's green) so that needs
looking at, but it will allow the camera app to start, take pictures and
close without crashing. Much of this is code taken from the emulator, but
some changes have been made and more are coming.